% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilled_lifestage.R
\name{add_cohort.chilled_lifestage}
\alias{add_cohort.chilled_lifestage}
\title{Add a cohort to a chilled_lifestage}
\usage{
\method{add_cohort}{chilled_lifestage}(x, number, cumDevelopment = 0, cumChill = 0)
}
\arguments{
\item{x}{A chilled_lifestage object}

\item{number}{(integer) The number of individuals in the cohort to be added}

\item{cumDevelopment}{(double) Their prior development (default = 0)}

\item{cumChill}{(double) Their prior chill accumulation (default = 0)}
}
\value{
The updated chilled_lifestage object
}
\description{
Add a cohort to a chilled_lifestage
}
\examples{
  eggs <- eggs |> add_cohort(100)

}
