% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{add_cohort.daylength_lifestage}
\alias{add_cohort.daylength_lifestage}
\title{Add a cohort to a daylength_lifestage}
\usage{
\method{add_cohort}{daylength_lifestage}(x, number)
}
\arguments{
\item{x}{A daylength_lifestage object}

\item{number}{The number of individuals in the cohort}
}
\value{
A modified daylength_lifestage object
}
\description{
Add a cohort to a daylength_lifestage
}
