% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{add_cohort.lifestage}
\alias{add_cohort.lifestage}
\title{Add a cohort to a lifestage}
\usage{
\method{add_cohort}{lifestage}(x, number, cumDevelopment = 0)
}
\arguments{
\item{x}{A lifestage object}

\item{number}{(integer) The number of individuals in the cohort to be added}

\item{cumDevelopment}{(double) Their prior development (default = 0)}
}
\value{
The updated lifestage
}
\description{
Add a cohort to a lifestage
}
\examples{
  larvae <- larvae |> add_cohort(100)

}
