% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{analytis2_development}
\alias{analytis2_development}
\title{Analytis 2 development model}
\usage{
analytis2_development(t, T0, Tu, a, b, c, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{T0}{The lower temperature threshold (°C)}

\item{Tu}{The upper temperature threshold (°C)}

\item{a}{A parameter controlling the height of the curve}

\item{b}{A parameter controlling skew of the curve}

\item{c}{A parameter controlling concavity of the curve}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520).
}
\examples{
plot(0:35, analytis2_development(0:35, T0 = 5, Tu = 33, a = 1, b = 5, c = 1), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
