% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{beta_development}
\alias{beta_development}
\title{Beta development model}
\usage{
beta_development(t, Rm, T0, Tm, Tu, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{Rm}{The maximum development rate}

\item{T0}{The lower temperature threshold (°C)}

\item{Tm}{The optimum temperature (°C)}

\item{Tu}{The upper temperature threshold (°C)}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520).
}
\examples{
plot(0:35, beta_development(0:35, Rm = 0.2, T0 = 5, Tm = 28, Tu = 33), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
