% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{briere2_development}
\alias{briere2_development}
\title{Briere 2 development model}
\usage{
briere2_development(t, T0, Tu, a, b, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{T0}{The lower temperature threshold (°C)}

\item{Tu}{The upper temperature threshold (°C)}

\item{a}{A parameter controlling the height of the curve}

\item{b}{A parameter controlling the shape of the curve}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Briere et al. (1999, Environ. Entomol. 28:22-29).
}
\examples{
plot(0:35, briere2_development(0:35, T0 = 5, Tu = 33, a = 0.0001, b = 4), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
