% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{calendar_development}
\alias{calendar_development}
\title{Calendar development model}
\usage{
calendar_development(t, Q, ...)
}
\arguments{
\item{t}{The temperature (°C) (NB not used)}

\item{Q}{The number of days required to complete the life stage}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Models development rate R (/d) as a simple accumulation of time.
}
\examples{
plot(1:30, calendar_development(1:30), type = "l", xlab = "Time required Q (d)", ylab = "Development rate R (/d)")

}
