% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_models.R
\name{cum_normal_variability}
\alias{cum_normal_variability}
\title{Cumulative normal variability}
\usage{
cum_normal_variability(D, sd, m = 1)
}
\arguments{
\item{D}{The developmental accumulation (1 = average required)}

\item{sd}{The standard deviation}

\item{m}{The median of the curve (default = 1)}
}
\value{
Probability of actual development
}
\description{
Cumulative normal variability
}
\examples{
plot(seq(0, 2, 0.01), cum_normal_variability(seq(0, 2, 0.01), sd = 0.1), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")

}
