% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilled_lifestage.R
\name{develop.chilled_lifestage}
\alias{develop.chilled_lifestage}
\title{Develop cohorts for one timestep}
\usage{
\method{develop}{chilled_lifestage}(x, drivers)
}
\arguments{
\item{x}{A chilled_lifestage object}

\item{drivers}{A tibble row with degreesC, daylength, daytrend and dayfraction}
}
\value{
The updated chilled_lifestage object
}
\description{
Develop cohorts for one timestep
}
\examples{
  eggs <- eggs |> develop(d)

}
