% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{develop.daylength_lifestage}
\alias{develop.daylength_lifestage}
\title{Develop daylength_lifestage for one timestep}
\usage{
\method{develop}{daylength_lifestage}(x, drivers)
}
\description{
Develop daylength_lifestage for one timestep
}
