% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{develop.lifestage}
\alias{develop.lifestage}
\title{Develop cohorts for one timestep}
\usage{
\method{develop}{lifestage}(x, drivers)
}
\arguments{
\item{x}{A lifestage object}

\item{drivers}{A tibble row with degreesC, daylength, daytrend and dayfraction}
}
\value{
The updated lifestage object
}
\description{
Develop cohorts for one timestep
}
\examples{
  larvae <- larvae |> develop(d)

}
