% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{eradicate_population}
\alias{eradicate_population}
\title{Remove all individuals across all lifestages}
\usage{
eradicate_population(pop)
}
\arguments{
\item{pop}{A list of lifestage objects}
}
\value{
Updated list of lifestage objects
}
\description{
Remove all individuals across all lifestages
}
