% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_constant_drivers}
\alias{example_constant_drivers}
\title{Make some driver data with constant temperature and daylength}
\usage{
example_constant_drivers(degC = 20, daylength = 14)
}
\arguments{
\item{degC}{The temperature (°C)}

\item{daylength}{The daylength (h)}
}
\value{
Driver data with constant temperature and daylength
}
\description{
Make some driver data with constant temperature and daylength
}
\examples{
View(example_constant_drivers())
plot_drivers(example_constant_drivers())

}
