% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_phenology_simulation}
\alias{example_phenology_simulation}
\title{Run an example phenology model and plot the results}
\usage{
example_phenology_simulation(length = 365)
}
\arguments{
\item{length}{The number of days to simulate for}
}
\value{
A plot of some example simulation results
}
\description{
Run an example phenology model and plot the results
}
