% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_timestep}
\alias{example_timestep}
\title{Example timestep}
\usage{
example_timestep(
  population,
  drivers,
  survival = c(eggs = 0.999, larvae = 0.99, pupae = 0.99, adults = 0.9),
  fecundity = 0.2
)
}
\arguments{
\item{population}{List of life stages}

\item{drivers}{Suitable temperature and daylength data for 1 time step}

\item{survival}{Stage survival parameters}

\item{fecundity}{Adult fecundity}
}
\value{
Updated population after 1 time step
}
\description{
Example timestep
}
