% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_varying_drivers}
\alias{example_varying_drivers}
\title{Make some driver data with varying temperature and daylength}
\usage{
example_varying_drivers(latitude = -37.44, degC = 15, length = 365)
}
\arguments{
\item{latitude}{The latitude of the location (for calculating daylengths)}

\item{degC}{The mean temperature}

\item{length}{The number of days returned}
}
\value{
Some example driver data with varying temperature and daylength
}
\description{
Make some driver data with varying temperature and daylength
}
\examples{
View(example_varying_drivers())
plot_drivers(example_varying_drivers())

}
