% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{expbase_development}
\alias{expbase_development}
\title{Exponential base development model}
\usage{
expbase_development(t, T0, sl, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{T0}{The base temperature for development (°C)}

\item{sl}{A slope parameter}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Exponential base development model
}
\examples{
plot(0:35, expbase_development(0:35, T0 = 10, sl = 0.01), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
