% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{exponential_development}
\alias{exponential_development}
\title{Exponential development model}
\usage{
exponential_development(t, sl, sp, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{sl}{A slope parameter}

\item{sp}{A spread parameter}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Exponential development model
}
\examples{
plot(0:35, exponential_development(0:35, sl = 0.01, sp = 0.1), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
