% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fit_day_degree_model}
\alias{fit_day_degree_model}
\title{Fit a linear day degree model to data}
\usage{
fit_day_degree_model(degC, days)
}
\arguments{
\item{degC}{A list of the temperatures}

\item{days}{A list of the required number of days for development}
}
\value{
Parameters for the base temperature (°C) and day degree requirement (°d)
}
\description{
Fit a linear day degree model to data
}
\examples{
fit_day_degree_model(c(10, 15, 20, 25, 30), c(50, 25, 12, 6, 4))

}
