% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{gaussian_development}
\alias{gaussian_development}
\title{Gaussian development model}
\usage{
gaussian_development(t, Rm, T0, Tm, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{Rm}{The maximum development rate}

\item{T0}{The approximate lower threshold temperature (°C) where R = 5\% of Rm}

\item{Tm}{The optimum temperature (°C)}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Also known as the Pradhan-Taylor model.
Here it has been modified to accommodate biologically meaningful input parameters.
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Pradhan (1945, Proc. Nat. Inst. Sci. India 11:73-80),
Han et al (2000, Annals Entomol. Soc. Am. 93:536-540),
Walgama & Zalucki (2006, Insect Sci. 13:109-118).
}
\examples{
plot(0:35, gaussian_development(0:35, Rm = 0.2, T0 = 10, Tm = 30), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
