% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{kill_all.daylength_lifestage}
\alias{kill_all.daylength_lifestage}
\title{Kill all individuals in the daylength_lifestage}
\usage{
\method{kill_all}{daylength_lifestage}(x)
}
\description{
Kill all individuals in the daylength_lifestage
}
