% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{lactin1_development}
\alias{lactin1_development}
\title{Lactin 1 development model}
\usage{
lactin1_development(t, rho, Tu, delta, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{rho}{A constant defining the development rate at the optimal temperature}

\item{Tu}{The upper temperature threshold (°C)}

\item{delta}{A range over which physiological breakdown becomes the overriding influence}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
A modified form of the Logan 1 model, to eliminate an unnecessary parameter.
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Lactin et al. (1995, Env. Entomol. 24: 68-75).
}
\examples{
plot(0:35, lactin1_development(0:35, rho = 0.15, Tu = 33, delta = 6.4), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
