% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{lactin2_development}
\alias{lactin2_development}
\title{Lactin 2 development model}
\usage{
lactin2_development(t, rho, Tu, delta, lambda, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{rho}{A constant defining the development rate at the optimal temperature}

\item{Tu}{The upper temperature threshold (°C)}

\item{delta}{A range over which physiological breakdown becomes the overriding influence}

\item{lambda}{A parameter that forces a low temperature threshold}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Lactin et al. (1995, Env. Entomol. 24: 68-75).
}
\examples{
plot(0:35, lactin2_development(0:35, rho = 0.15, Tu = 33, delta = 6.4, lambda = -0.3), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
