% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilling_models.R
\name{linear_chill}
\alias{linear_chill}
\title{Linear chill function}
\usage{
linear_chill(t, T_chill, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{T_chill}{The chill threshold temperature (°C)}
}
\value{
The chill accumulation
}
\description{
Linear chill function
}
