% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{linear_development}
\alias{linear_development}
\title{Linear development model}
\usage{
linear_development(t, b, q, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{b}{The base temperature for development (°C)}

\item{q}{The total day-degree requirement (°d)}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Models linear insect development rate as a function of temperature.
}
\examples{
plot(0:35, linear_development(0:35, b = 10, q = 100), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
