% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{logan1_development}
\alias{logan1_development}
\title{Logan 1 development model}
\usage{
logan1_development(t, psi, rho, Tu, delta, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{psi}{A scalar for the development rate}

\item{rho}{A constant defining the development rate at the optimal temperature}

\item{Tu}{The upper temperature threshold (°C)}

\item{delta}{A temperature range over which physiological breakdown becomes the overriding influence (°C)}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Also known as the Logan-6 model.
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Logan et al. (1976, Env. Entomol. 5:1133-1140),
Logan (1988, Env. Entomol. 17: 359-376),
Lactin et al. (1995, Env. Entomol. 24: 68-75).
}
\examples{
plot(0:35, logan1_development(0:35, psi = 0.05, rho = 0.19, Tu = 33, delta = 5), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
