% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{logan2_development}
\alias{logan2_development}
\title{Logan 2 development model}
\usage{
logan2_development(t, a, K, rho, Tu, delta, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{a}{A scalar for the development rate}

\item{K}{An empirical constant}

\item{rho}{A constant defining the development rate at the optimal temperature}

\item{Tu}{The upper temperature threshold (°C)}

\item{delta}{The temperature range over which physiological breakdown becomes the overriding influence (°C)}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Also known as the Logan-10 model.
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Logan et al. (1976, Env. Entomol. 5:1133-1140),
Logan (1988, Env. Entomol. 17: 359-376).
}
\examples{
plot(0:35, logan2_development(0:35, a = 0.5, K = 50, rho = 0.19, Tu = 33, delta = 2), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
