% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_models.R
\name{logistic_variability}
\alias{logistic_variability}
\title{Logistic variability}
\usage{
logistic_variability(D, w, m = 1)
}
\arguments{
\item{D}{The developmental accumulation}

\item{w}{The width of the curve, equivalent to 3*sd. At x = m - w, y = 0.01 and at x = m + w, y = 0.99}

\item{m}{The mean of the curve (default = 1)}
}
\value{
Probability of actual development
}
\description{
Logistic variability
}
\examples{
plot(seq(0, 2, 0.01), logistic_variability(seq(0, 2, 0.01), w = 0.5), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")
plot(seq(6, 18, 0.01), logistic_variability(seq(6, 18, 0.01), w = 1, m = 14), type = "l", xlab = "Daylength (h)", ylab = "Proportion breaking diapause")

}
