% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{maturing.daylength_lifestage}
\alias{maturing.daylength_lifestage}
\title{Get total maturing from daylength_lifestage in the current timestep}
\usage{
\method{maturing}{daylength_lifestage}(x, ...)
}
\description{
Get total maturing from daylength_lifestage in the current timestep
}
