% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{maturing.lifestage}
\alias{maturing.lifestage}
\title{Get total maturing from lifestage in the current timestep}
\usage{
\method{maturing}{lifestage}(x, ...)
}
\arguments{
\item{x}{A lifestage object}
}
\value{
The number maturing from this lifestage in the current timestep
}
\description{
Get total maturing from lifestage in the current timestep
}
\examples{
  maturing(larvae)

}
