% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilled_lifestage.R
\name{new_chilled_lifestage}
\alias{new_chilled_lifestage}
\title{Constructor}
\usage{
new_chilled_lifestage(
  name,
  devFunction,
  devParameters = list(),
  varFunction,
  varParameters = list(),
  chillFunction,
  chillParameters = list(),
  chillResponse
)
}
\arguments{
\item{name}{Name of the stage e.g. "Eggs"}

\item{devFunction}{The development function to use e.g. linear_development}

\item{devParameters}{Named list of parameters for the development function e.g. c(b = 10, q = 500)}

\item{varFunction}{The variability function to use e.g. cum_normal_variability}

\item{varParameters}{Named list of parameters for the variability function e.g. c(sd = 0.1)}

\item{chillFunction}{The chill accumulation function e.g. linear_chill}

\item{chillParameters}{Named list of parameters for the variability function e.g. c(T_chill = 5)}

\item{chillResponse}{The function relating the cumulative chill to the relative development rate (0 to 1).}
}
\value{
A chilled_lifestage object
}
\description{
Constructor
}
\examples{
eggs = new_chilled_lifestage(
  name = "Eggs",
  devFunction = linear_development,
  devParameters = c(b = 7.4, q = 428),
  varFunction = cum_normal_variability,
  varParameters = c(sd = 0.1),
  chillFunction = linear_chill,
  chillParameters = c(T_chill = 13.9),
  chillResponse = function(x) 1 - 0.46 * exp(-0.0005 * x)
)

}
