% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{new_daylength_lifestage}
\alias{new_daylength_lifestage}
\title{Constructor for daylength diapausing life stage}
\usage{
new_daylength_lifestage(name, Threshold, ReqTrend, Variation)
}
\arguments{
\item{name}{The name of the stage e.g. "Diapausing eggs"}

\item{Threshold}{The required daylength (in hours) for completing the stage (positive for increasing daylength, negative for decreasing)}

\item{ReqTrend}{The required trend in daylength: 1 for increasing daylength, -1 for decreasing daylength}

\item{Variation}{The variation around the threshold (in hours)}
}
\value{
A daylength_lifestage object
}
\description{
Constructor for daylength diapausing life stage
}
