% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologysim_package.R
\docType{package}
\name{phenologysim-package}
\alias{phenologysim}
\alias{phenologysim-package}
\title{phenologysim: A framework for phenology simulation}
\description{
Tools for phenology modelling.
}
\author{
\strong{Maintainer}: John Kean \email{john.kean@agresearch.co.nz}

Authors:
\itemize{
  \item John Kean \email{john.kean@agresearch.co.nz}
}

}
\keyword{internal}
