% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{plot_daily_drivers}
\alias{plot_daily_drivers}
\title{Plot daily drivers}
\usage{
plot_daily_drivers(driver_data)
}
\arguments{
\item{driver_data}{A data frame with columns for date, dayfraction, degreesC, daylength and daytrend}
}
\value{
A ggplot graph showing daily Tmin, Tmax, Tmean and daylength
}
\description{
Plot daily drivers
}
