% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{plot_drivers}
\alias{plot_drivers}
\title{Plot drivers}
\usage{
plot_drivers(driver_data)
}
\arguments{
\item{driver_data}{A data frame with columns for date, dayfraction, degreesC, daylength and daytrend}
}
\value{
A ggplot graph
}
\description{
Plot drivers
}
\examples{
plot_drivers(example_varying_drivers())

}
