% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot_population_trajectory}
\alias{plot_population_trajectory}
\title{Plot a population trajectory}
\usage{
plot_population_trajectory(population_data, type = "area", x_label = "")
}
\arguments{
\item{population_data}{Data with the first column for the x-axis and remaining columns being life stages to be plotted}

\item{type}{The plot type: either "area" (default) or "line"}

\item{x_label}{The label for the x-axis}
}
\value{
A plot of the simulation results
}
\description{
Plot a population trajectory
}
