% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{prepare_daylengths}
\alias{prepare_daylengths}
\title{Set up daylength data frame for phenology modelling}
\usage{
prepare_daylengths(data, latitude = NULL)
}
\arguments{
\item{data}{A data frame with columns for date and daylength}
}
\value{
A data frame with columns for date, daylength and daytrend (1 if daylength is increasing, -1 if it is decreasing)
}
\description{
Set up daylength data frame for phenology modelling
}
