% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{prepare_temperatures}
\alias{prepare_temperatures}
\title{Set up temperature data frame for phenology modelling}
\usage{
prepare_temperatures(data)
}
\arguments{
\item{data}{A data frame with columns for date and optionally Tmin, Tmax and/or Tmean}
}
\value{
A data frame with columns for date, dayfraction and degreesC
}
\description{
Set up temperature data frame for phenology modelling
}
