% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{print}
\alias{print}
\title{Print method for lifestage objects}
\usage{
print(x, ...)
}
\arguments{
\item{x}{A lifestage object}
}
\description{
Print method for lifestage objects
}
\examples{
 print(larvae)
}
