% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_models.R
\name{regniere1_variability}
\alias{regniere1_variability}
\title{Regniere 1 variability}
\usage{
regniere1_variability(D, a, b, m = 1)
}
\arguments{
\item{D}{The developmental accumulation}

\item{a}{A parameter determining the steepness of the curve}

\item{b}{A parameter determining the skew}

\item{m}{The median of the curve (default = 1)}
}
\value{
Probability of actual development
}
\description{
See Régnière (1984, Can. Ento. 116:1367-1376).
}
\examples{
plot(seq(0, 2, 0.01), regniere1_variability(seq(0, 2, 0.01), a = 20, b = 2), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")

}
