% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_models.R
\name{regniere2_variability}
\alias{regniere2_variability}
\title{Regniere 2 variability}
\usage{
regniere2_variability(D, a, b, m = 1.073)
}
\arguments{
\item{D}{The developmental accumulation}

\item{a}{A parameter determining the steepness of the curve}

\item{b}{A parameter determining the skew}

\item{m}{A parameter determining the position of the curve (default = 1.073)}
}
\value{
Probability of actual development
}
\description{
See Régnière (1984, Can. Ento. 116:1367-1376).
}
\examples{
plot(seq(0, 2, 0.01), regniere2_variability(seq(0, 2, 0.01), a = 15, b = 2), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")

}
