% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rescale_dd}
\alias{rescale_dd}
\title{Rescale degree day requirement}
\usage{
rescale_dd(oldReq, oldBase, newBase = 10, pivot = 20)
}
\arguments{
\item{oldReq}{The old day degree requirement (°d)}

\item{oldBase}{The old base temperature (°C)}

\item{newBase}{The desired new base temperature (°C) - default is 10 °C}

\item{pivot}{The pivot temperature (°C) - the default is 25 °C}
}
\value{
The adjusted day degree requirement (°d)
}
\description{
Rescale degree day requirement
}
\examples{
rescale_dd(100, 12)

}
