% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation (generic handler)}
\usage{
run_simulation(population, initialise, timestep, drivers, verbose = FALSE, ...)
}
\arguments{
\item{population}{(function)  function that creates a list of population life stage objects}

\item{initialise}{(list)      initial values for non-zero life stages}

\item{timestep}{(function)  function used for the timestep}

\item{drivers}{(dataframe) with columns for date, dayfraction, degreesC, daylength and daytrend}

\item{verbose}{(boolean)   whether to print progress}

\item{...}{additional parameters required by the timestep function}
}
\value{
(dataframe) drivers with JulianDate, days_passed and life stage totals added as new columns
}
\description{
Run a simulation (generic handler)
}
