% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_models.R
\name{sigmoid_development}
\alias{sigmoid_development}
\title{Sigmoid development model}
\usage{
sigmoid_development(t, Rm, T0, Tm, ...)
}
\arguments{
\item{t}{The temperature (°C)}

\item{Rm}{The upper limit of the sigmoid line (/d)}

\item{T0}{The approximate lower threshold temperature (°C) where R = 5\% of Rm}

\item{Tm}{The approximate optimal temperature (°C) where R = 95\% of Rm}

\item{...}{Unused variables}
}
\value{
The proportion of total development requirement that is completed today (/d)
}
\description{
Also known as the Davidson or logistic model.
Here it has been modified to accommodate biologically meaningful input parameters.
See Mirhosseini et al. (2017, Annals Entomol. Soc. Am. 110:507-520),
Casagrande et al. (1987, Environ. Entomol. 16:556-562).
}
\examples{
plot(0:35, sigmoid_development(0:35, Rm = 0.1, T0 = 10, Tm = 30), type = "l", xlab = "Temperature T (°C)", ylab = "Development rate R (/d)")

}
