% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_models.R
\name{step_variability}
\alias{step_variability}
\title{Step variability}
\usage{
step_variability(D, m = 1)
}
\arguments{
\item{D}{The developmental accumulation}

\item{m}{The median of the curve (default = 1)}
}
\value{
Probability of actual development
}
\description{
Step variability
}
\examples{
plot(seq(0, 2, 0.01), step_variability(seq(0, 2, 0.01)), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")
plot(seq(0, 2, 0.01), step_variability(seq(0, 2, 0.01), m = 0.5), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")

}
