% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{survive.lifestage}
\alias{survive.lifestage}
\title{Apply survival to all cohorts in the lifestage}
\usage{
\method{survive}{lifestage}(x, drivers, survival)
}
\arguments{
\item{x}{A lifestage object}

\item{drivers}{Tibble including the dayfraction}

\item{survival}{Survival rate /day}
}
\value{
The updated lifestage object
}
\description{
Apply survival to all cohorts in the lifestage
}
\examples{
  larvae <- larvae |> survive(0.99, d)

}
