% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{total.daylength_lifestage}
\alias{total.daylength_lifestage}
\title{Get total number in daylength_lifestage}
\usage{
\method{total}{daylength_lifestage}(x, ...)
}
\description{
Get total number in daylength_lifestage
}
