% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{total_population}
\alias{total_population}
\title{Return total population size across all lifestages}
\usage{
total_population(pop)
}
\arguments{
\item{pop}{A list of lifestage objects}
}
\value{
Numeric sum of all total(lifestage)
}
\description{
Return total population size across all lifestages
}
