% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_models.R
\name{weibull_variability}
\alias{weibull_variability}
\title{Weibull variability}
\usage{
weibull_variability(D, a = 0, b = NA, c = NA, m = 1)
}
\arguments{
\item{D}{The developmental accumulation}

\item{a}{The minimum developmental accumulation required (default = 0)}

\item{b}{The scale parameter (width of the curve) (default = NA)}

\item{c}{The shape of the curve (default = NA)}

\item{m}{The median of the curve (default = 1)}
}
\value{
Probability of actual development
}
\description{
This is flexible implementation. Parameters may be specified as (b,c), (a,b,c), (a,b,m) or (a,c,m)
}
\examples{
plot(seq(0, 2, 0.01), weibull_variability(seq(0, 2, 0.01), a = 0.5, c = 3.39), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")
plot(seq(0, 10, 0.01), weibull_variability(seq(0, 10, 0.01), a = 4, b = 2.2, c = 4), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")
plot(seq(0, 10, 0.01), weibull_variability(seq(0, 10, 0.01), a = 4, b = 2.2, m = 6), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")
plot(seq(0, 10, 0.01), weibull_variability(seq(0, 10, 0.01), a = 4, c = 4, m = 6), type = "l", xlab = "Developmental accumulation", ylab = "Probability of development")

}
