% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilled_lifestage.R
\name{develop.chilled_lifestage}
\alias{develop.chilled_lifestage}
\title{Develop cohorts for one timestep}
\usage{
\method{develop}{chilled_lifestage}(x, drivers)
}
\description{
Develop cohorts for one timestep
}
