% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{develop.daylength_lifestage}
\alias{develop.daylength_lifestage}
\title{Develop daylength_lifestage for one timestep}
\usage{
\method{develop}{daylength_lifestage}(x, drivers)
}
\arguments{
\item{x}{A daylength_lifestage object}

\item{drivers}{The temperature and daylength drivers}
}
\value{
A modified daylength_lifestage object
}
\description{
Develop daylength_lifestage for one timestep
}
