% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{kill_all.lifestage}
\alias{kill_all.lifestage}
\title{Kill all individuals in the lifestage}
\usage{
\method{kill_all}{lifestage}(x)
}
\arguments{
\item{x}{A lifestage object}
}
\value{
The updated lifestage object
}
\description{
Kill all individuals in the lifestage
}
\examples{
  larvae <- larvae |> kill_all()

}
