% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{new_lifestage}
\alias{new_lifestage}
\title{Constructor for lifestage object}
\usage{
new_lifestage(
  name,
  devFunction,
  devParameters = list(),
  varFunction,
  varParameters = list()
)
}
\arguments{
\item{name}{Name of the stage e.g. "First instar larvae"}

\item{devFunction}{Name of the development function to use e.g. "linear_development"}

\item{devParameters}{Named list of parameters for the development function e.g. c(b = 10, q = 500)}

\item{varFunction}{Name of the variability function to use e.g. "cum_normal_variability"}

\item{varParameters}{Named list of parameters for the variability function e.g. c(sd = 0.1)}
}
\value{
A lifestage object
}
\description{
Constructor for lifestage object
}
\examples{
larvae = new_lifestage(
  name = "Larvae",
  devFunction = linear_development,
  devParameters = list(b = 10, q = 200),
  varFunction = cum_normal_variability,
  varParameters = list(sd = 0.1)
)

}
