% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{prepare_drivers}
\alias{prepare_drivers}
\title{Set up driver data for phenology modelling}
\usage{
prepare_drivers(data, ...)
}
\arguments{
\item{data}{A data frame with columns for date, temperature and optionally daylength}

\item{latitude}{Optional alternative if daylength is not supplied}
}
\value{
A data frame with columns for date, dayfraction, degreesC, daylength and daytrend
}
\description{
This is a shorthand for prepare_temperatures() then prepare_daylengths().
}
