% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{print.daylength_lifestage}
\alias{print.daylength_lifestage}
\title{Print method for daylength_lifestage objects}
\usage{
\method{print}{daylength_lifestage}(x)
}
\description{
Print method for daylength_lifestage objects
}
