% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{print.lifestage}
\alias{print.lifestage}
\title{Print method for lifestage objects}
\usage{
\method{print}{lifestage}(x)
}
\arguments{
\item{x}{A lifestage object}
}
\description{
Print method for lifestage objects
}
\examples{
 print(larvae)

}
