% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation (generic handler)}
\usage{
run_simulation(population, timestep, drivers, verbose = FALSE, ...)
}
\arguments{
\item{population}{A list of population life stage objects}

\item{timestep}{The function used for the timestep}

\item{drivers}{A dataframe with columns for date, dayfraction, degreesC, daylength and daytrend}

\item{verbose}{Boolean whether to print progress}

\item{...}{Additional parameters required by the timestep function}
}
\value{
A dataframe with drivers, JulianDate, days_passed and life stage totals
}
\description{
Run a simulation (generic handler)
}
\examples{
results <- run_simulation(
  population = example_lifestages,
  timestep   = example_timestep,
  drivers    = driver_data
)

}
