% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{survive.daylength_lifestage}
\alias{survive.daylength_lifestage}
\title{Apply survival to a daylength_lifestage}
\usage{
\method{survive}{daylength_lifestage}(x, drivers, survival)
}
\description{
Apply survival to a daylength_lifestage
}
