% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{survive.lifestage}
\alias{survive.lifestage}
\title{Apply survival to all cohorts in the lifestage}
\usage{
\method{survive}{lifestage}(x, survival_rate, day_fraction = 1)
}
\arguments{
\item{x}{A lifestage object}

\item{survival_rate}{Survival rate /day}

\item{day_fraction}{Proportion of the day that this applies to (default = 1)}
}
\value{
The updated lifestage object
}
\description{
Apply survival to all cohorts in the lifestage
}
\examples{
  larvae <- larvae |> survive(0.99, 0.25)

}
