% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength_lifestage.R
\name{total.daylength_lifestage}
\alias{total.daylength_lifestage}
\title{Get total number in daylength_lifestage}
\usage{
\method{total}{daylength_lifestage}(x, ...)
}
\arguments{
\item{x}{A daylength_lifestage object}
}
\value{
The total number of individuals in the lifestage
}
\description{
Get total number in daylength_lifestage
}
