% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage.R
\name{total.lifestage}
\alias{total.lifestage}
\title{Get total number in lifestage}
\usage{
\method{total}{lifestage}(x, ...)
}
\value{
The total number of individuals in that lifestage, across all cohorts
}
\description{
Get total number in lifestage
}
\examples{
  total(larvae)

}
